Camera zoom and positioning + Walls Top/Bottom Screen
Camera can zoom in and out between 2 public params (MaxHeight and MinHeight), swapping from one to the other as quick as MultiplierDistanceHeight variable is (high value = quicker zooms).

A wall appears on top / bottom of the screen if one of the player reaches the limit (plus an extra).
For debugging, the wall is visible. To set it invisible, search for the SetVisibility in the blueprints, and set it to false.

This version works when 2 players are on the same height. If not, walls that appear could push the players when moving. (To be improved)