Camera Screen Edges without walls
Now we dont spawn invisible walls to avoid players leaving the screen. We transfomr its positions to screen positions and check how far are to the Edges (setted with a public variable). If they are near the edge, we teleport back the player that is moving (now with a hardcoded value) .

Things to improve:
- What happens if 2 players are trying to leave the screen moving in opposite directions.
- Small shake when moving into edge in 2 directions.